tech(api): move answer model
