+++ api/lib/infrastructure/serializers/json/scoring-simulator/scoring-simulation-dataset-serializer.js
@@ -1,6 +1,6 @@
 import { ScoringSimulation } from '../../../../domain/models/ScoringSimulation.js';
 import { ScoringSimulationDataset } from '../../../../domain/models/ScoringSimulationDataset.js';
-import { Answer } from '../../../../domain/models/Answer.js';
+import { Answer } from '../../../../../src/evaluation/domain/models/Answer.js';
 
 const deserialize = function (request) {
   const simulations = request.payload.dataset.simulations.map(
