+++ api/src/evaluation/infrastructure/repositories/answer-repository.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import jsYaml from 'js-yaml';
 import { knex } from '../../../../db/knex-database-connection.js';
 import { ChallengeAlreadyAnsweredError, NotFoundError } from '../../../../lib/domain/errors.js';
-import { Answer } from '../../../../lib/domain/models/Answer.js';
+import { Answer } from '../../domain/models/Answer.js';
 import * as answerStatusDatabaseAdapter from '../../../../lib/infrastructure/adapters/answer-status-database-adapter.js';
 
 function _adaptAnswerToDb(answer) {
