+++ api/src/shared/domain/models/Assessment.js
@@ -1,5 +1,5 @@
 import { ObjectValidationError } from '../../../../lib/domain/errors.js';
-import { Answer } from '../../../../lib/domain/models/Answer.js';
+import { Answer } from '../../../evaluation/domain/models/Answer.js';
 import { CertificationChallengeLiveAlertStatus } from '../../../../lib/domain/models/CertificationChallengeLiveAlert.js';
 
 const courseIdMessage = {
