+++ api/tests/evaluation/unit/domain/models/Answer_test.js
@@ -1,6 +1,6 @@
-import { Answer } from '../../../../lib/domain/models/Answer.js';
-import { Skill } from '../../../../lib/domain/models/Skill.js';
-import { domainBuilder, expect } from '../../../test-helper.js';
+import { Answer } from '../../../../../src/evaluation/domain/models/Answer.js';
+import { Skill } from '../../../../../lib/domain/models/Skill.js';
+import { domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('Unit | Domain | Models | Answer', function () {
   describe('constructor', function () {
