+++ api/tests/unit/domain/models/Examiner_test.js
@@ -1,4 +1,4 @@
-import { Answer } from '../../../../lib/domain/models/Answer.js';
+import { Answer } from '../../../../src/evaluation/domain/models/Answer.js';
 import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
 import { Examiner } from '../../../../lib/domain/models/Examiner.js';
 import { expect, domainBuilder, sinon } from '../../../test-helper.js';
