+++ api/tests/unit/domain/usecases/get-correction-for-answer_test.js
@@ -1,6 +1,6 @@
 import { getCorrectionForAnswer } from '../../../../lib/domain/usecases/get-correction-for-answer.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
-import { Answer } from '../../../../lib/domain/models/Answer.js';
+import { Answer } from '../../../../src/evaluation/domain/models/Answer.js';
 import { AssessmentNotCompletedError, NotFoundError } from '../../../../lib/domain/errors.js';
 import { expect, sinon, catchErr, domainBuilder } from '../../../test-helper.js';
 import { LearningContentResourceNotFound } from '../../../../lib/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
